+++ src/main/java/gift/exception/BadRequestExceptions/UserNotFoundException.java
@@ -1,0 +1,6 @@
+package gift.exception.BadRequestExceptions;
+
+public class UserNotFoundException extends BadRequestException {
+    public UserNotFoundException() {super();}
+    public UserNotFoundException(String message) {super(message);}
+}
